backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
django-filter>=24.0

